+++ docs/changelog.rst
@@ -7,7 +7,7 @@ Versions are year-based with a strict backwards-compatibility policy.
 The third digit is only for regressions.
 
 
-14.1.0 (2014-08-21)
+14.1.0 (2014-08-22)
 -------------------